refactor(fødselsnummer): Endrer begrep fra fødselsnummer til personidentifikator
Felles for vanlige norske fødselsnummere er at de er personidentifikator. I forbindelse med
innføring av ny personidentifikator bør vi også distansere oss litt fra fødselsnummerbegrepet.

BREAKING CHANGE: Endret navn på modul fødselsnummer til personidentifikator. Noen funksjoner har
også fått nye navn.
