+++ lib/navfaker.ts
@@ -1,18 +1,18 @@
 import Dato from './dato/dato';
-import Fødselsnummer from './fodselsnummer/fodselsnummer';
+import PersonIdentifikator from './fodselsnummer/personIdentifikator';
 import Person from './person/person';
 import Random from './random/random';
 import Telefon from './telefon/telefon';
 
 class NavFaker {
     public random: Random;
-    public fødselsnummer: Fødselsnummer;
+    public personIdentifikator: PersonIdentifikator;
     public dato: Dato;
     public person: Person;
     public telefon: Telefon;
 
     constructor() {
-        this.fødselsnummer = new Fødselsnummer(this);
+        this.personIdentifikator = new PersonIdentifikator(this);
         this.random = new Random();
         this.dato = new Dato(this);
         this.person = new Person(this);
